KAM-1826: cards to its own component from home
