+++ apps/consultation-portal/screens/Home/Home.json
@@ -4,9 +4,6 @@
       "title": "Öll mál",
       "description": "Samráðsgátt greiðir leið hagsmunaaðila, almennings og félagasamtaka að opinberum samráðsferlum sem mögulegt er að taka þátt í.",
       "keywords": "Opinbert samráð, lagasetning, frumvarp, reglugerð, stefna, umsögn, umsagnir"
-    },
-    "card": {
-      "eyebrowText": "Umsagnir"
     }
   },
   "heroBanner": {
@@ -74,5 +71,8 @@
       "noOptions": "Enginn stofnun",
       "placeholder": "Veldu stofnun"
     }
+  },
+  "card": {
+    "eyebrowText": "Umsagnir"
   }
 }
