+++ apps/consultation-portal/screens/Home/Home.tsx
@@ -1,13 +1,4 @@
-import {
-  Box,
-  GridColumn,
-  GridContainer,
-  GridRow,
-  Tiles,
-  Text,
-  Stack,
-  LoadingDots,
-} from '@island.is/island-ui/core'
+import { GridColumn, GridContainer, GridRow } from '@island.is/island-ui/core'
 import {
   HeroBanner,
   MobileFilter,
@@ -15,23 +6,14 @@ import {
   SearchAndFilter,
 } from './components/'
 import localization from './Home.json'
-import {
-  ArrOfStatistics,
-  ArrOfTypes,
-  Case,
-  CaseFilter,
-} from '../../types/interfaces'
-import { Card, EmptyState, Pagination, Layout } from '../../components'
+import { ArrOfTypes, CaseFilter } from '../../types/interfaces'
+import { Layout } from '../../components'
 import {
   useFetchStatistics,
   useFrontPageFilters,
   useIsMobile,
 } from '../../hooks'
-import {
-  CARDS_PER_PAGE,
-  FILTERS_FRONT_PAGE_KEY,
-} from '../../utils/consts/consts'
-import { MapCaseStatuses } from '../../types/enums'
+import Cards from './components/Cards/Cards'
 
 interface HomeProps {
   types: ArrOfTypes
@@ -58,73 +40,6 @@ export const Index = ({ types }: HomeProps) => {
   })
 
   const { statistics } = useFetchStatistics()
-
-  const renderCards = () => {
-    if (getCasesLoading) {
-      return (
-        <Box
-          display="flex"
-          width="full"
-          alignItems="center"
-          justifyContent="center"
-          style={{ height: 200 }}
-        >
-          <LoadingDots color="blue" large />
-        </Box>
-      )
-    }
-
-    if (cases?.length === 0) {
-      return <EmptyState isCase />
-    }
-
-    return (
-      <>
-        {cases && (
-          <Tiles space={3} columns={[1, 1, 1, 2, 3]}>
-            {cases.map((item: Case, index: number) => {
-              const card = {
-                id: item.id,
-                title: item.name,
-                tag: MapCaseStatuses[item.statusName],
-                published: item.created,
-                processEnds: item.processEnds,
-                processBegins: item.processBegins,
-                eyebrows: [item.typeName, item.institutionName],
-                caseNumber: item.caseNumber,
-              }
-              return (
-                <Card key={index} card={card} frontPage showPublished>
-                  <Stack space={2}>
-                    <Text variant="eyebrow" color="purple400">
-                      {`${loc.card.eyebrowText}: ${item.adviceCount}`}
-                    </Text>
-                    <Box
-                      style={{
-                        wordBreak: 'break-word',
-                        height: '105px',
-                      }}
-                      overflow="hidden"
-                    >
-                      <Text variant="small" color="dark400">
-                        {item.shortDescription}
-                      </Text>
-                    </Box>
-                  </Stack>
-                </Card>
-              )
-            })}
-          </Tiles>
-        )}
-        <Pagination
-          filters={filters}
-          setFilters={(arr: CaseFilter) => setFilters(arr)}
-          totalPages={Math.ceil(total / CARDS_PER_PAGE)}
-          localStorageId={FILTERS_FRONT_PAGE_KEY}
-        />
-      </>
-    )
-  }
 
   return (
     <Layout
@@ -168,7 +83,13 @@ export const Index = ({ types }: HomeProps) => {
             )}
           </GridColumn>
           <GridColumn span={['12/12', '12/12', '12/12', '9/12', '9/12']}>
-            {renderCards()}
+            <Cards
+              getCasesLoading={getCasesLoading}
+              cases={cases}
+              total={total}
+              filters={filters}
+              setFilters={setFilters}
+            />
           </GridColumn>
         </GridRow>
       </GridContainer>
